:truck: rename: entityForJpa -> entity로 경로 변경
